+++ magpylib/_src/display/display_matplotlib.py
@@ -1,15 +1,24 @@
 """ matplotlib draw-functionalities"""
 
+from itertools import cycle
 import numpy as np
+import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
 from magpylib._src.default_classes import default_settings as Config
-from magpylib._src.display.disp_utility import (
+from magpylib._src.display.display_utility import (
     get_rot_pos_from_path,
     draw_arrow_from_vertices,
     draw_arrowed_circle,
     place_and_orient_model3d,
+    faces_cuboid,
+    faces_cylinder,
+    system_size,
+    faces_sphere,
+    faces_cylinder_section,
 )
-
+from magpylib._src.input_checks import check_excitations
+from magpylib._src.style import get_style
+from magpylib._src.display.display_utility import MagpyMarkers
 
 def draw_directs_faced(faced_objects, colors, ax, show_path, size_direction):
     """draw direction of magetization of faced magnets
@@ -308,3 +317,183 @@ def draw_model3d_extra(obj, style, show_path, ax, color):
             kwargs.update({k: v for k, v in tr.items() if k not in ("type", "args")})
             args = tr.get("args", [])
             getattr(ax, tr["type"])(*args, **kwargs)
+
+
+def display_matplotlib(
+    *obj_list_semi_flat,
+    axis=None,
+    markers=None,
+    path=True,
+    zoom=0,
+    color_sequence=None,
+    **kwargs,
+):
+    """
+    Display objects and paths graphically with the matplotlib backend.
+
+    - axis: matplotlib axis3d object
+    - markers: list of marker positions
+    - path: bool / int / list of ints
+    - zoom: zoom level, 0=tight boundaries
+    - color_sequence: list of colors for object coloring
+    """
+    # pylint: disable=protected-access
+    # pylint: disable=too-many-branches
+    # pylint: disable=too-many-statements
+
+    # apply config default values if None
+    if color_sequence is None:
+        color_sequence = Config.display.colorsequence
+    # create or set plotting axis
+    if axis is None:
+        fig = plt.figure(dpi=80, figsize=(8, 8))
+        ax = fig.add_subplot(111, projection="3d")
+        ax.set_box_aspect((1, 1, 1))
+        generate_output = True
+    else:
+        ax = axis
+        generate_output = False
+
+    # draw objects and evaluate system size --------------------------------------
+
+    # draw faced objects and store vertices
+    points = []
+    dipoles_color = []
+    sensors_color = []
+    faced_objects_color = []
+
+    for semi_flat_obj, color in zip(obj_list_semi_flat, cycle(color_sequence)):
+        flat_objs = [semi_flat_obj]
+        if getattr(semi_flat_obj, "children", None) is not None:
+            flat_objs = semi_flat_obj.children
+            if getattr(semi_flat_obj, "position", None) is not None:
+                flat_objs += [semi_flat_obj]
+                color = color if semi_flat_obj.style.color is None else semi_flat_obj.style.color
+        for obj in flat_objs:
+            style = get_style(obj, Config, **kwargs)
+            color = style.color if style.color is not None else color
+            lw = 0.25
+            faces = None
+            if obj.style.model3d.extra:
+                draw_model3d_extra(obj, style, path, ax, color)
+            if obj.style.model3d.show:
+                if obj._object_type == "Cuboid":
+                    lw = 0.5
+                    faces = faces_cuboid(obj, path)
+                elif obj._object_type == "Cylinder":
+                    faces = faces_cylinder(obj, path)
+                elif obj._object_type == "CylinderSegment":
+                    faces = faces_cylinder_section(obj, path)
+                elif obj._object_type == "Sphere":
+                    faces = faces_sphere(obj, path)
+                elif obj._object_type == "Line":
+                    if style.arrow.show:
+                        check_excitations([obj])
+                    arrow_size = style.arrow.size if style.arrow.show else 0
+                    arrow_width = style.arrow.width
+                    points += draw_line(
+                        [obj], path, color, arrow_size, arrow_width, ax
+                    )
+                elif obj._object_type == "Loop":
+                    if style.arrow.show:
+                        check_excitations([obj])
+                    arrow_width = style.arrow.width
+                    arrow_size = style.arrow.size if style.arrow.show else 0
+                    points += draw_circular(
+                        [obj], path, color, arrow_size, arrow_width, ax
+                    )
+                elif obj._object_type == "Sensor":
+                    sensors_color += [color]
+                    points += draw_pixel(
+                        [obj],
+                        ax,
+                        color,
+                        style.pixel.color,
+                        style.pixel.size,
+                        style.pixel.symbol,
+                        path,
+                    )
+                elif obj._object_type == "Dipole":
+                    dipoles_color += [color]
+                    points += [obj.position]
+                elif obj._object_type == "CustomSource":
+                    draw_markers(np.array([obj.position]), ax, color, symbol="*", size=10)
+                    name = (
+                        obj.style.name
+                        if obj.style.name is not None
+                        else str(type(obj).__name__)
+                    )
+                    ax.text(*obj.position, name, horizontalalignment="center")
+                    points += [obj.position]
+                if faces is not None:
+                    faced_objects_color += [color]
+                    alpha = style.opacity
+                    pts = draw_faces(faces, color, lw, alpha, ax)
+                    points += [np.vstack(pts).reshape(-1, 3)]
+                    if style.magnetization.show:
+                        check_excitations([obj])
+                        pts = draw_directs_faced(
+                            [obj], [color], ax, path, style.magnetization.size
+                        )
+                        points += pts
+            if path:
+                marker, line = style.path.marker, style.path.line
+                if style.path.show:
+                    points += draw_path(
+                        obj,
+                        color,
+                        marker.symbol,
+                        marker.size,
+                        marker.color,
+                        line.style,
+                        line.width,
+                        ax,
+                    )
+
+    # markers -------------------------------------------------------
+    if markers is not None and markers:
+        m = MagpyMarkers()
+        style = get_style(m, Config, **kwargs)
+        markers = np.array(markers)
+        s = style.marker
+        draw_markers(markers, ax, s.color, s.symbol, s.size)
+        points += [markers]
+
+    # draw direction arrows (based on src size) -------------------------
+    # objects with faces
+
+    # determine system size -----------------------------------------
+    limx1, limx0, limy1, limy0, limz1, limz0 = system_size(points)
+
+    # make sure ranges are not null
+    limits = np.array([[limx0, limx1], [limy0, limy1], [limz0, limz1]])
+    limits[np.squeeze(np.diff(limits)) == 0] += np.array([-1, 1])
+    sys_size = np.max(np.diff(limits))
+    c = limits.mean(axis=1)
+    m = sys_size.max() / 2
+    ranges = np.array([c - m * (1 + zoom), c + m * (1 + zoom)]).T
+
+    # draw all system sized based quantities -------------------------
+    # sensors
+    sensors = [obj for obj in obj_list_semi_flat if obj._object_type == "Sensor"]
+
+    # dipoles
+    dipoles = [obj for obj in obj_list_semi_flat if obj._object_type == "Dipole"]
+
+    # not optimal for loop if many sensors/dipoles
+    for sensor in sensors:
+        style = get_style(sensor, Config, **kwargs)
+        draw_sensors([sensor], ax, sys_size, path, style.size)
+    for dipole, color in zip(dipoles, dipoles_color):
+        style = get_style(dipole, Config, **kwargs)
+        draw_dipoles([dipole], ax, sys_size, path, style.size, color, style.pivot)
+
+    # plot styling --------------------------------------------------
+    ax.set(
+        **{f"{k}label": f"{k} [mm]" for k in "xyz"},
+        **{f"{k}lim": r for k, r in zip("xyz", ranges)},
+    )
+
+    # generate output ------------------------------------------------
+    if generate_output:
+        plt.show()
