+++ magpylib/_src/display/display_utility.py
