+++ magpylib/_src/display/plotly_sensor_mesh.py
